requirements.txt
instabot==0.79.0
python-dotenv==0.10.3
tqdm==4.40.0
